Fix name bug in DIX recipes
When Recipe Robot generated the DIX recipes, it created variable
references the package name during processing. When I changed the name from "Disk
Inventory X" to "DiskInventoryX" in a recipe override, this caused the versioner
procesor to complain that the info.plist it was looking for couldn't be found
since the app's name is Disk Inventory X.app instead of DiskInventoryX.app.

This commit hardcodes the name "Disk Inventory X.app" in all internal processing
so users are free to rename the package without causing a problem.